+++ e2e/tests/solana/set_root.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/gagliardetto/solana-go"
+	"github.com/smartcontractkit/chainlink-ccip/chains/solana/gobindings/timelock"
 
 	"github.com/smartcontractkit/mcms"
 	"github.com/smartcontractkit/mcms/sdk"
@@ -35,19 +36,27 @@ func (s *SolanaTestSuite) Test_Solana_SetRoot() {
 	auth, err := solana.PrivateKeyFromBase58(privateKey)
 	s.Require().NoError(err)
 
+	metadata, err := solanasdk.NewChainMetadata(
+		0,
+		s.MCMProgramID,
+		testPDASeedSetRootTest,
+		s.Roles[timelock.Proposer_Role].AccessController.PublicKey(),
+		s.Roles[timelock.Canceller_Role].AccessController.PublicKey(),
+		s.Roles[timelock.Bypasser_Role].AccessController.PublicKey())
+	s.Require().NoError(err)
 	validUntil := time.Now().Add(10 * time.Hour).Unix()
 	proposal, err := mcms.NewProposalBuilder().
 		SetVersion("v1").
 		SetValidUntil(uint32(validUntil)).
 		SetDescription("proposal to test SetRoot").
 		SetOverridePreviousRoot(true).
-		AddChainMetadata(s.ChainSelector, types.ChainMetadata{MCMAddress: mcmAddress}).
+		AddChainMetadata(s.ChainSelector, metadata).
 		AddOperation(types.Operation{
 			ChainSelector: s.ChainSelector,
 			Transaction: types.Transaction{
 				To:               recipientAddress,
 				Data:             []byte("0x"),
-				AdditionalFields: json.RawMessage(`{"value": 0}`), // FIXME: not used right now; check with jonghyeon.park
+				AdditionalFields: json.RawMessage(`{"value": 0, "accounts": []}`), // FIXME: not used right now; check with jonghyeon.park
 			},
 		}).
 		Build()