+++ factory.go
@@ -42,12 +42,9 @@ func newEncoder(
 	return encoder, nil
 }
 
-// newTimelockConverterFromExecutor returns a new TimelockConverter that can convert timelock proposals
+// newTimelockConverter a new TimelockConverter that can convert timelock proposals
 // for the given chain.
-func newTimelockConverterFromExecutor(
-	csel types.ChainSelector,
-	executor sdk.TimelockExecutor,
-) (sdk.TimelockConverter, error) {
+func newTimelockConverter(csel types.ChainSelector) (sdk.TimelockConverter, error) {
 	family, err := types.GetChainSelectorFamily(csel)
 	if err != nil {
 		return nil, err
@@ -58,14 +55,9 @@ func newTimelockConverterFromExecutor(
 		return &evm.TimelockConverter{}, nil
 
 	case cselectors.FamilySolana:
-		solanaExecutor, ok := executor.(*solana.TimelockExecutor)
-		if !ok {
-			return nil, fmt.Errorf("unable to cast sdk executor to solana TimelockExecutor")
-		}
-
-		return solana.NewTimelockConverter(solanaExecutor.Client()), nil
+		return &solana.TimelockConverter{}, nil
 
 	default:
-		return nil, fmt.Errorf("unsupported executor type: %T", executor)
+		return nil, fmt.Errorf("unsupported chain family %s", family)
 	}
 }