+++ proposal.go
@@ -156,6 +156,15 @@ func (p *Proposal) Validate() error {
 		return err
 	}
 
+	// Validate chain metadata for each chain selector
+	// Should only be needed for timelock proposals (specifically solana proposals),
+	// but this might change as new chain families are added
+	for chainSelector, metadata := range p.ChainMetadata {
+		if err := validateChainMetadata(metadata, chainSelector); err != nil {
+			return err
+		}
+	}
+
 	// Validate all chains in operations have an entry in chain metadata
 	for _, op := range p.Operations {
 		if _, ok := p.ChainMetadata[op.ChainSelector]; !ok {
@@ -165,7 +174,7 @@ func (p *Proposal) Validate() error {
 
 	for _, op := range p.Operations {
 		// Chain specific validations.
-		if err := ValidateAdditionalFields(op.Transaction.AdditionalFields, op.ChainSelector); err != nil {
+		if err := validateAdditionalFields(op.Transaction.AdditionalFields, op.ChainSelector); err != nil {
 			return err
 		}
 	}